Audio-Video-Match/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
ca-certificates=2020.12.8=haa95532_0
certifi=2020.12.5=py38haa95532_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=py38_0
freetype=2.10.4=hd328e21_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
intel-openmp=2020.2=254
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.0=py38hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
lz4-c=1.9.2=hf4a77e7_3
matplotlib=3.3.2=haa95532_0
matplotlib-base=3.3.2=py38hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.2.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
ninja=1.10.2=py38h6d14046_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
olefile=0.46=py_0
openssl=1.1.1i=h2bbff1b_0
pillow=8.0.1=py38h4fa10fc_0
pip=20.3.3=py38haa95532_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38ha925a31_4
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=py_0
pytorch=1.7.1=py3.8_cuda101_cudnn7_0
qt=5.9.7=vc14h73c81de_0
scipy=1.5.2=py38h14eb087_0
setuptools=51.0.0=py38haa95532_2
sip=4.19.13=py38ha925a31_0
six=1.15.0=py38haa95532_0
sqlite=3.33.0=h2a8f88b_0
tk=8.6.10=he774522_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu101
tornado=6.1=py38h2bbff1b_0
typing_extensions=3.7.4.3=py_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
